backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# OpenAI
openai>=1.66.5,<2.0.0
openai-agents==0.0.6
tiktoken==0.8.0

# Rate Limiting
slowapi==0.1.9
limits==3.14.0

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# HTTP
httpx==0.28.1
anyio==4.7.0

# Utilities
python-dotenv==1.0.1
orjson==3.10.12
structlog==24.4.0

# Testing (dev)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
